feat: fix decryptor module
